+++ cpp/Configuration.h
@@ -1,6 +1,11 @@
+#pragma once
 
-// #define ANDROID_APP_THEME   
-// #define ANDROID_APP_ACCENT
-// #define ANDROID_APP_FORGROUND
-// #define ANDROID_APP_BACKGROUND
-#define ANDROID_MATERIAL_THEME_PRIMARY_COLOR             "#103e69"
+// -------------------------------------------------------------------------- ANDROID Material Colors definitions
+// Material Theme = Light (to change it go to the `MainLib.qml` and update the `Material.theme: Light` to `Dark` or `System`)
+// https://doc.qt.io/qt-5/qtquickcontrols2-material.html
+#define ANDROID_MATERIAL_ACCENT                     "#103e69"
+#define ANDROID_MATERIAL_PRIMARY                    "#103e69"
+#define ANDROID_MATERIAL_PRIMARY_FOREGROUND         "#fff"
+#define ANDROID_MATERIAL_FOREGROUND                 ""          // not used yet anywhere
+#define ANDROID_MATERIAL_BACKGROUND                 ""          // not used yet anywhere
+// --------------------------------------------------------------------------------------------------------------