+++ qml/MainLib.qml
@@ -16,9 +16,10 @@ Page {
     property bool embedVideoQML: false
     property bool callAutoAccept: false
     property var videoItem
-    Material.primary: GlobalObject.materialPrimaryColor()
-    // Material.theme: Material.System
-    // Material.accent: Material.Purple
+    
+    Material.theme: Material.Light
+    Material.primary: GlobalObject.materialColors().primary
+    Material.accent: GlobalObject.materialColors().accent
     StackView {
         id: stack
         anchors.fill: parent