+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 
+## [0.0.22]
+### Changed
+
+* Adding Material Accent and Foreground colors to Configuration.h.
+* Remove extra space between slide menu and top menu.
+* Force Active focus on TextField in sending message after each send.
+
+
 ## [0.0.21]
 ### Changed